+++ configs/settings.py
@@ -1,0 +1,4 @@
+"""Application configuration settings."""
+
+# Placeholder for configuration variables
+
