+++ make.ps1
@@ -1,0 +1,3 @@
+# PowerShell build script placeholder
+
+Write-Host "Build script not yet implemented"
