+++ src/backtest/__init__.py
@@ -1,0 +1,1 @@
+"""Backtesting package."""
