+++ src/backtest/engine.py
@@ -1,0 +1,4 @@
+"""Backtest engine placeholder."""
+
+# TODO: Implement backtesting engine
+
