+++ src/data_fetch.py
@@ -1,0 +1,4 @@
+"""Utilities for downloading and caching raw market data."""
+
+# TODO: Implement data retrieval logic
+
