+++ src/live/__init__.py
@@ -1,0 +1,1 @@
+"""Live trading package."""
