+++ src/ml/feature_engineering.py
@@ -1,0 +1,4 @@
+"""Feature engineering utilities."""
+
+# TODO: Implement feature transformation functions
+
