+++ src/ml/train_cli.py
@@ -1,0 +1,4 @@
+"""Command-line interface for model training."""
+
+# TODO: Implement CLI argument parsing and training invocation
+
