+++ src/utils/env.py
@@ -1,0 +1,4 @@
+"""Environment helpers for configuration and credentials."""
+
+# TODO: Add environment handling utilities
+
